sending output to pipe

+++ roslib2/src/roslib2/distro.py
@@ -330,7 +330,10 @@ class Distro(object):
                 # code.ros.org is a big hack.
                 tmp_dir = tempfile.mkdtemp()
                 tmp_distro_file = os.path.join(tmp_dir, os.path.split(source_uri)[-1])
-                subprocess.check_call(['svn','export',source_uri,tmp_distro_file])
+                p = subprocess.Popen(['svn','export',source_uri,tmp_distro_file], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+                p.communicate()
+                if p.returncode != 0:
+                    raise Exception("svn export [%s] failed"%(source_uri))
                 with open(tmp_distro_file) as f:
                     y = yaml.load(f.read())
                 shutil.rmtree(tmp_dir)
